// ui-catalogue/src/components/molecules/CuentasResultados/types.tsx
export interface Ii18nCuentasResultados {
  'CuentaResultados:detailTbVentasNetas': string
  'CuentaResultados:labelTbEjercicio1': string
  'CuentaResultados:labelTbEjercicio2': string
  'CuentaResultados:labelTbEjercicio3': string
  'CuentaResultados:detailTbVariacionExistencias': string
  'CuentaResultados:detailTbTrabajosInmovilizado': string
  'CuentaResultados:detailTbOtrosIngresos': string
  'CuentaResultados:detailTbValorProducción': string
  'CuentaResultados:detailTbAprovisionamientos': string
  'CuentaResultados:detailTbMateriasPrimas': string
  'CuentaResultados:detailTbSubcontragos': string
  'CuentaResultados:detailTbBrutoExplotación': string
  'CuentaResultados:detailTbGastosOperativos': string
  'CuentaResultados:detailTbDotaciónInmovilizado': string
  'CuentaResultados:detailTbProvTrafico': string
  'CuentaResultados:detailTbOtrosGastos': string
  'CuentaResultados:detailTbCostesReorganiczacion': string
  'CuentaResultados:detailTbNetoExplotación': string
  'CuentaResultados:detailTbGastosFinancieros': string
  'CuentaResultados:detailTbInteresesSimilares': string
  'CuentaResultados:detailTbPerdidasInversiones': string
  'CuentaResultados:detailTbOtrosGastosFinancieros': string
  'CuentaResultados:detailTbIngresosFinancieros': string
  'CuentaResultados:detailTbIngresosInmoFinanciero': string
  'CuentaResultados:detailTbIngresosGrupo': string
  'CuentaResultados:detailTbIngresosInversores': string
  'CuentaResultados:detailTbOrdinario': string
  'CuentaResultados:detailTbImpuestos': string
  'CuentaResultados:detailTbImpuestosAntex': string
  'CuentaResultados:detailTbIngresosExtraordinarios': string
  'CuentaResultados:detailTbGastosExtraordinarios': string
  'CuentaResultados:detailTbExtraordinarios': string
  'CuentaResultados:detailTbImpuestoExtraordinarios': string
  'CuentaResultados:detailTbBeneficios': string
  'CuentaResultados:detailTbParticipacionAccionistas': string
  'CuentaResultados:detailTbParticipacionEmpleados': string
  'CuentaResultados:detailTbAjusteIngresos': string
  'CuentaResultados:detailTbTercerasPartes': string
  'CuentaResultados:detailTbBeneficioNeto': string
  'CuentaResultados:detailTbPerdidasNetas': string
  'CuentaResultados:detailTbExportaciones': string
  'CuentaResultados:detailTbDistribIniciales': string
  'CuentaResultados:detailTbIngresosNetos': string
  'CuentaResultados:detailTbOtrasAcciones': string
  'CuentaResultados:detailTbPerdidasNetasBis': string
  'CuentaResultados:detailTbDividendos': string
  'CuentaResultados:detailTbRetenciones': string
  'CuentaResultados:detailTbOtrasDeducciones': string
  'CuentaResultados:detailTbBeneficiosDistribuidos': string
  'CuentaResultados:detailTbSalariosDirectores': string
}

export interface ICuentasResultadosRequestForm {
  ventasNetas1: string
  ventasNetas2: string
  ventasNetas3: string
  variacionExistencias1: string
  variacionExistencias2: string
  variacionExistencias3: string
  trabajosInmovilizado1: string
  trabajosInmovilizado2: string
  trabajosInmovilizado3: string
  otrosIngresos1: string
  otrosIngresos2: string
  otrosIngresos3: string
  valorProducción1: string
  valorProducción2: string
  valorProducción3: string
  aprovisionamientos1: string
  aprovisionamientos2: string
  aprovisionamientos3: string
  materiasPrimas1: string
  materiasPrimas2: string
  materiasPrimas3: string
  subcontragos1: string
  subcontragos2: string
  subcontragos3: string
  brutoExplotación1: string
  brutoExplotación2: string
  brutoExplotación3: string
  gastosOperativos1: string
  gastosOperativos2: string
  gastosOperativos3: string
  dotaciónInmovilizado1: string
  dotaciónInmovilizado2: string
  dotaciónInmovilizado3: string
  provTrafico1: string
  provTrafico2: string
  provTrafico3: string
  otrosGastos1: string
  otrosGastos2: string
  otrosGastos3: string
  costesReorganiczacion1: string
  costesReorganiczacion2: string
  costesReorganiczacion3: string
  netoExplotación1: string
  netoExplotación2: string
  netoExplotación3: string
  gastosFinancieros1: string
  gastosFinancieros2: string
  gastosFinancieros3: string
  interesesSimilares1: string
  interesesSimilares2: string
  interesesSimilares3: string
  perdidasInversiones1: string
  perdidasInversiones2: string
  perdidasInversiones3: string
  otrosGastosFinancieros1: string
  otrosGastosFinancieros2: string
  otrosGastosFinancieros3: string
  ingresosFinancieros1: string
  ingresosFinancieros2: string
  ingresosFinancieros3: string
  ingresosInmoFinanciero1: string
  ingresosInmoFinanciero2: string
  ingresosInmoFinanciero3: string
  ingresosGrupo1: string
  ingresosGrupo2: string
  ingresosGrupo3: string
  ingresosInversores1: string
  ingresosInversores2: string
  ingresosInversores3: string
  ordinario1: string
  ordinario2: string
  ordinario3: string
  impuestos1: string
  impuestos2: string
  impuestos3: string
  impuestosAntex1: string
  impuestosAntex2: string
  impuestosAntex3: string
  ingresosExtraordinarios1: string
  ingresosExtraordinarios2: string
  ingresosExtraordinarios3: string
  gastosExtraordinarios1: string
  gastosExtraordinarios2: string
  gastosExtraordinarios3: string
  extraordinarios1: string
  extraordinarios2: string
  extraordinarios3: string
  impuestoExtraordinarios1: string
  impuestoExtraordinarios2: string
  impuestoExtraordinarios3: string
  beneficios1: string
  beneficios2: string
  beneficios3: string
  participacionAccionistas1: string
  participacionAccionistas2: string
  participacionAccionistas3: string
  participacionEmpleados1: string
  participacionEmpleados2: string
  participacionEmpleados3: string
  ajusteIngresos1: string
  ajusteIngresos2: string
  ajusteIngresos3: string
  tercerasPartes1: string
  tercerasPartes2: string
  tercerasPartes3: string
  beneficioNeto1: string
  beneficioNeto2: string
  beneficioNeto3: string
  perdidasNetas1: string
  perdidasNetas2: string
  perdidasNetas3: string
  exportaciones1: string
  exportaciones2: string
  exportaciones3: string
  distribIniciales1: string
  distribIniciales2: string
  distribIniciales3: string
  ingresosNetos1: string
  ingresosNetos2: string
  ingresosNetos3: string
  otrasAcciones1: string
  otrasAcciones2: string
  otrasAcciones3: string
  perdidasNetasBis1: string
  perdidasNetasBis2: string
  perdidasNetasBis3: string
  dividendos1: string
  dividendos2: string
  dividendos3: string
  retenciones1: string
  retenciones2: string
  retenciones3: string
  otrasDeducciones1: string
  otrasDeducciones2: string
  otrasDeducciones3: string
  beneficiosDistribuidos1: string
  beneficiosDistribuidos2: string
  beneficiosDistribuidos3: string
  salariosDirectores1: string
  salariosDirectores2: string
  salariosDirectores3: string
}

export interface ICuentasResultados {
  i18n: Ii18nCuentasResultados
  data: ICuentasResultadosRequestForm
}